renamed deletePostHandlers to deletePostHelpers
